Question Page Refactoring
